requirements.txt
admin-logs==0.2.14
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
Django==5.2
django-cors-headers==4.7.0
django-extensions==4.1
django-grappelli==4.0.1
django-jazzmin==3.0.1
django-js-asset==3.1.2
django-picklefield==3.3
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
graphviz==0.20.3
kombu==5.5.3
pillow==11.2.1
prompt_toolkit==3.0.51
pydotplus==2.0.2
PyJWT==2.9.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
